requirements_2_clean.txt
absl-py==2.3.1
albumentations==1.1.0
archspec==0.2.3
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
av==16.0.1
beautifulsoup4==4.12.3
boltons==24.0.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
#clip==1.0
clip @ git+https://github.com/openai/CLIP.git
colorama==0.4.6
#conda==24.9.2
#conda-build==24.9.0
#conda_index==0.5.0
#conda-libmamba-solver==24.9.0
#conda-package-handling==2.4.0
#conda_package_streaming==0.11.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dlib==20.0.0
dnspython==2.7.0
efficientnet_pytorch==0.7.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
expecttest==0.2.1
filelock==3.16.1
filterpy==1.4.5
fonttools==4.61.1
frozendict==2.4.6
fsspec==2024.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
grpcio==1.76.0
h2==4.1.0
hf-xet==1.2.0
hpack==4.0.0
huggingface-hub==0.36.0
hyperframe==6.0.1
hypothesis==6.115.5
idna==3.10
ImageIO==2.37.2
imgaug==0.4.0
importlib_resources==6.4.5
imutils==0.5.4
iopath==0.1.10
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
libarchive-c==5.1
#libmambapy==1.5.10
lief==0.14.1
lintrunner==0.12.5
lmdb==1.7.5
loralib==0.1.2
#mamba==1.5.10
Markdown==3.10
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
#menuinst==2.1.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
#nvidia-cublas-cu12==12.4.5.8
#nvidia-cuda-cupti-cu12==12.4.127
#nvidia-cuda-nvrtc-cu12==12.4.127
#nvidia-cuda-runtime-cu12==12.4.127
#nvidia-cudnn-cu12==9.1.0.70
#nvidia-cufft-cu12==11.2.1.3
#nvidia-curand-cu12==10.3.5.147
#nvidia-cusolver-cu12==11.6.1.9
#nvidia-cusparse-cu12==12.3.1.170
#nvidia-nccl-cu12==2.21.5
#nvidia-nvjitlink-cu12==12.4.127
#nvidia-nvtx-cu12==12.4.127
nvidia-cublas-cu12>=12.6.0
nvidia-cuda-cupti-cu12>=12.6.0
nvidia-cuda-nvrtc-cu12>=12.6.0
nvidia-cuda-runtime-cu12>=12.6.0
nvidia-cudnn-cu12>=9.3.0
nvidia-cufft-cu12>=11.2.0
nvidia-curand-cu12>=10.3.0
nvidia-cusolver-cu12>=11.6.0
nvidia-cusparse-cu12>=12.5.0
nvidia-nccl-cu12>=2.21.5
nvidia-nvjitlink-cu12>=12.6.0
nvidia-nvtx-cu12>=12.6.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
optree==0.13.0
packaging==24.1
pandas==2.3.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==25.3
pkginfo==1.11.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.2.0
prettytable==3.17.0
prompt_toolkit==3.0.48
protobuf==6.33.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
#pycosat==0.6.6
pycparser==2.22
Pygments==2.18.0
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytorchvideo==0.1.5
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
referencing==0.35.1
regex==2025.11.3
requests==2.32.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
shapely==2.1.2
simplejson==3.20.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
#sympy==1.13.1
sympy>=1.13.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.12.20
timm==1.0.22
tokenizers==0.22.1
#torch==2.5.1+cu124
torch==2.9.1
#torchaudio==2.5.1+cu124
#torchaudio==2.6.1 --index-url https://download.pytorch.org/whl/cu121
torchelastic==0.2.2
torchtoolbox==0.1.8.2
#torchvision==0.20.1+cu124
#torchvision==0.21.1 --index-url https://download.pytorch.org/whl/cu121
tqdm==4.66.5
traitlets==5.14.3
transformers==4.57.3
#triton==3.1.0
triton>=3.2.0
truststore==0.9.2
types-dataclasses==0.6.6
typing_extensions==4.12.2
tzdata==2025.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.4
wheel==0.44.0
yacs==0.1.8
zipp==3.20.2
zstandard==0.23.0
